refactor code review: comments and lint fix
