+++ packages/static-cs-lite/lib/util/imageFrame/convert/color/convertRGBColorByPlane.js
@@ -1,6 +1,13 @@
-/* eslint-disable no-plusplus */
-const assertArrayDivisibility = require("./assertArrayDivisibility");
+/* eslint-disable no-plusplus, no-param-reassign */
+const assertArrayDivisibility = require("../../../assertArrayDivisibility");
 
+/**
+ * Convert pixel data with RGB (by plane) Photometric Interpretation to RGBA
+ *
+ * @param {ImageFrame} imageFrame
+ * @param {Uint8ClampedArray} rgbaBuffer buffer result (this param is mutate)
+ * @returns {void}
+ */
 function converter(imageFrame, rgbaBuffer) {
   if (!assertArrayDivisibility(imageFrame, 3, ["decodeRGB: rgbBuffer must not be undefined", "decodeRGB: rgbBuffer length must be divisble by 3"])) {
     return;
