+++ packages/static-cs-lite/lib/util/index.js
@@ -1,0 +1,5 @@
+const imageFrameUtils = require("./imageFrame");
+const assertArrayDivisibility = require("./assertArrayDivisibility");
+
+exports.imageFrameUtils = imageFrameUtils;
+exports.assertArrayDivisibility = assertArrayDivisibility;
